Extract Importer concern from app/services/
This is one of a set of refactors the H3 developers discussed on 03-19-2025, each of which will be submitted as a separate PR.
