+++ app/importers/import_error.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class ImportError < StandardError
+end
