--- app/services/importers/error.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-module Importers
-  class Error < StandardError
-  end
-end
